Modify shebang and add quotes to bash/sh scripts as needed.
Modified the shebang to use env, this allow people to use whatever
path they like for bash,sh or rc.  Added quotes to variables and
expressions as needed in bash/sh scripts to avoid quotting issues
in some environment and in some versions of bash/sh this missing
quotes caused strange errors. Hopefully this changes fixes that.